requirements.txt
dvc==3.23.0
numpy==1.24.4
torch==1.13.1
coverage==7.3.1
mysql-connector-python==8.1.0
hvac==1.2.1
kafka-python==2.0.2